backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.30
fastf1>=3.3.0
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
langchain-core>=0.1.0
sse-starlette>=1.8.0
